Fase 5 (0): el mundo pasa a 384 de alto, de y = −64 a 319
- Coordenadas reales de Minecraft 1.18+: el terreno de siempre queda igual de 0 hacia arriba;
  debajo, piedra hasta el lecho de roca en −64, cuevas más hondas con lava desde −54 y los
  minerales de Minecraft bajando hasta el fondo. El vacío empieza en −128.
- Índices de columna con la fila 0 en MIN_Y (blockIndex, indexY); claves de posición con y de
  9 bits; ediciones guardadas con índice de 24 bits y ediciones en la red con y de 16 bits
  (protocolo 5).
- Migración automática de los mundos guardados (DO, IndexedDB y memoria): ediciones y cofres,
  hornos, fogatas y carteles, una sola vez.
- Cliente: mallador y luz con 384 filas (origen del chunk en MIN_Y), mapa de lluvia en flotante,
  rayos, colisiones y colocación con los nuevos límites.
- Pruebas: tests/height.test.ts (terreno bajo cero, ediciones, migración, mallado).

+++ src/client/world/worldWorker.ts
@@ -9,7 +9,7 @@ export type WorkerRequest =
   | { type: 'spawn'; id: number };
 
 export type WorkerResponse =
-  | { type: 'gen'; id: number; cx: number; cz: number; blocks: Uint16Array; tint: Uint8Array; heights: Uint8Array }
+  | { type: 'gen'; id: number; cx: number; cz: number; blocks: Uint16Array; tint: Uint8Array; heights: Int16Array }
   | {
       type: 'mesh';
       id: number;
